WorkoutPlan/serializers.py
from .models import WorkoutPlan
from rest_framework import serializers


class WorkoutPlanSerializer(serializers.ModelSerializer):
    class Meta:
        model = WorkoutPlan
        fields = '__all__'


